# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Casenova API Documentation",
        "description": "Casenova API Documentation",
        "contact": {
            "email": "admin@casenova.uz"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://casenova.lc/api/v1",
            "description": "Api Server"
        }
    ],
    "paths": {
...
        "/equipment": {
            "get": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Get list of equipments",
                "description": "Returns list of equipments",
                "operationId": "getEquipments",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Create new equipment",
                "description": "Returns equipment data",
                "operationId": "storeEquipments",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEquipmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/equipment/{id}": {
            "get": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Get equipment information",
                "description": "Returns equipment data",
                "operationId": "getEquipmentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Equipment id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EquipmentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Update equipment",
                "description": "Returns equipment data",
                "operationId": "updateEquipment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Equipment id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEquipmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/equipment/{equipment}/image": {
            "post": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Update image for equipment",
                "description": "Returns equipment status",
                "operationId": "imageEquipment",
                "parameters": [
                    {
                        "name": "equipment",
                        "in": "path",
                        "description": "Equipment id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "$ref": "#/components/schemas/BaseModel/properties/property_uploads"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseModel/properties/property_boolean_result"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreEquipmentRequest": {
                "title": "StoreEquipmentRequest",
                "properties": {
                    "ru": {
                        "$ref": "#/components/schemas/StoreEquipmentTranslationRequest"
                    },
                    "uz": {
                        "$ref": "#/components/schemas/StoreEquipmentTranslationRequest"
                    },
                    "en": {
                        "$ref": "#/components/schemas/StoreEquipmentTranslationRequest"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreEquipmentRequest"
                }
            },
            "StoreEquipmentTranslationRequest": {
                "title": "StoreEquipmentTranslationRequest",
                "properties": {
                    "name": {
                        "$ref": "#/components/schemas/EquipmentTranslation/properties/name"
                    },
                    "title": {
                        "$ref": "#/components/schemas/EquipmentTranslation/properties/title"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreEquipmentTranslationRequest"
                }
            },
...
            "UpdateEquipmentRequest": {
                "title": "UpdateEquipmentRequest",
                "properties": {
                    "ru": {
                        "$ref": "#/components/schemas/UpdateEquipmentTranslationRequest"
                    },
                    "uz": {
                        "$ref": "#/components/schemas/UpdateEquipmentTranslationRequest"
                    },
                    "en": {
                        "$ref": "#/components/schemas/UpdateEquipmentTranslationRequest"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateEquipmentRequest"
                }
            },
            "UpdateEquipmentTranslationRequest": {
                "title": "UpdateEquipmentTranslationRequest",
                "properties": {
                    "name": {
                        "$ref": "#/components/schemas/EquipmentTranslation/properties/name"
                    },
                    "title": {
                        "$ref": "#/components/schemas/EquipmentTranslation/properties/title"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateEquipmentTranslationRequest"
                }
            },
...
            "EquipmentCollection": {
                "title": "EquipmentCollection",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EquipmentResource"
                        }
                    }
                },
                "type": "object"
            },
            "EquipmentResource": {
                "title": "EquipmentResource",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/Equipment/properties/id"
                    },
                    "image": {
                        "$ref": "#/components/schemas/Equipment/properties/image"
                    },
                    "createdAt": {
                        "$ref": "#/components/schemas/Equipment/properties/created_at"
                    },
                    "updatedAt": {
                        "$ref": "#/components/schemas/Equipment/properties/updated_at"
                    },
                    "translations": {
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/EquipmentTranslationResource"
                                }
                            ]
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EquipmentResource"
                }
            },
...
            "EquipmentTranslationResource": {
                "title": "EquipmentTranslationResource",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/EquipmentTranslation/properties/id"
                    },
                    "equipmentId": {
                        "$ref": "#/components/schemas/EquipmentTranslation/properties/equipment_id"
                    },
                    "locale": {
                        "$ref": "#/components/schemas/EquipmentTranslation/properties/locale"
                    },
                    "name": {
                        "$ref": "#/components/schemas/EquipmentTranslation/properties/name"
                    },
                    "title": {
                        "$ref": "#/components/schemas/EquipmentTranslation/properties/title"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EquipmentTranslationResource"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}